extract reading selected bestuursorgaan from context, rename madaten to mandaatOptions
