+++ app/components/mandatenbeheer/mandaat-bestuursorganen-selector.hbs
@@ -8,7 +8,7 @@
       @allowClear={{@allowClear}}
       @disabled={{@disabled}}
       @placeholder="Zoek mandaat"
-      @options={{this.mandaten}}
+      @options={{this.mandaatOptions}}
       @selected={{this.mandaat}}
       @searchField="bestuursfunctie.label"
       @onChange={{this.select}}
