--- app/utils/bestuursorgaan-meta-ttl.js
@@ -1,12 +1,0 @@
-export const getBestuursorgaanMetaTtl = (bestuursorgaan) => {
-  if (!bestuursorgaan) {
-    return;
-  }
-  const bestuursorgaanUri = bestuursorgaan.uri;
-
-  return `
-    @prefix ext: <http://mu.semte.ch/vocabularies/ext/> .
-
-    ext:applicationContext ext:currentBestuursorgaan <${bestuursorgaanUri}> .
-  `;
-};
